+++ ts/project.test.ts
@@ -1,17 +1,12 @@
-import { Check } from "./check.ts";
-import { Executable } from "./executable.ts";
-import { Plugin, Project } from "./project.ts";
 import { describe, it } from "https://deno.land/std@0.206.0/testing/bdd.ts";
 import * as garn from "./mod.ts";
-import * as nix from "./nix.ts";
 import { assertStdout, assertSuccess, runExecutable } from "./testUtils.ts";
-import { Package } from "./mod.ts";
 
-const assertTypeIsCheck = (_c: Check) => {};
-const assertTypeIsExecutable = (_e: Executable) => {};
-const assertTypeIsPackage = (_p: Package) => {};
+const assertTypeIsCheck = (_c: garn.Check) => {};
+const assertTypeIsExecutable = (_e: garn.Executable) => {};
+const assertTypeIsPackage = (_p: garn.Package) => {};
 
-const _testTypeCheckingOfAddCheck = (project: Project) => {
+const _testTypeCheckingOfAddCheck = (project: garn.Project) => {
   const p = project
     .addExecutable("unrelated1", "true")
     .addCheck("check", "true")
@@ -23,7 +18,7 @@ const _testTypeCheckingOfAddCheck = (project: Project) => {
   p.check``;
 };
 
-const _testTypeCheckingOfAddCheckTemplate = (project: Project) => {
+const _testTypeCheckingOfAddCheckTemplate = (project: garn.Project) => {
   const p = project
     .addExecutable("unrelated1", "true")
     .addCheck("check")`true`.addExecutable("unrelated2", "true");
@@ -34,7 +29,7 @@ const _testTypeCheckingOfAddCheckTemplate = (project: Project) => {
   p.check``;
 };
 
-const _testTypeCheckingOfAddExecutable = (project: Project) => {
+const _testTypeCheckingOfAddExecutable = (project: garn.Project) => {
   const p = project
     .addExecutable("unrelated1", "true")
     .addExecutable("shell", "true")
@@ -46,7 +41,7 @@ const _testTypeCheckingOfAddExecutable = (project: Project) => {
   p.shell``;
 };
 
-const _testTypeCheckingOfAddExecutableTemplate = (project: Project) => {
+const _testTypeCheckingOfAddExecutableTemplate = (project: garn.Project) => {
   const p = project
     .addExecutable("unrelated1", "true")
     .addExecutable("shell")`true`.addExecutable("unrelated2", "true");
@@ -61,7 +56,7 @@ describe("Project.add", () => {
   it("allows adding fields with .add", () => {
     const project = garn
       .mkProject({ description: "" }, {})
-      .add((self) => ({ ...self, foo: garn.shell("echo foo") }));
+      .add((_self) => ({ foo: garn.shell("echo foo") }));
     const output = runExecutable(project.foo);
     assertSuccess(output);
     assertStdout(output, "foo\n");
@@ -73,8 +68,8 @@ describe("Project.add", () => {
         { description: "", defaultEnvironment: garn.emptyEnvironment },
         {},
       )
-      .withDevTools([garn.mkPackage(nix.nixRaw`pkgs.hello`, "")])
-      .add((self) => ({ ...self, foo: self.shell("hello") }));
+      .withDevTools([garn.mkPackage(garn.nix.nixRaw`pkgs.hello`, "")])
+      .add((self) => ({ foo: self.shell("hello") }));
     const output = runExecutable(project.foo);
     assertSuccess(output);
     assertStdout(output, "Hello, world!\n");
@@ -92,7 +87,7 @@ describe("Project.add", () => {
           `,
         },
       )
-      .add((self) => ({ ...self, foo: self.shell`${self.package}/bin/main` }));
+      .add((self) => ({ foo: self.shell`${self.package}/bin/main` }));
     const output = runExecutable(project.foo);
     assertSuccess(output);
     assertStdout(output, "main executable\n");
@@ -137,8 +132,8 @@ describe("Project.add", () => {
   });
 
   it("provides a nice type synonym for plugins that add a field", () => {
-    const plugin: Plugin<{ addedField: garn.Package }> = (p) => ({
-      addedField: garn.build``,
+    const plugin: garn.Plugin<{ addedField: garn.Package }> = (_p) => ({
+      addedField: garn.build(""),
     });
     const project = garn
       .mkProject(
@@ -152,8 +147,10 @@ describe("Project.add", () => {
   });
 
   it("provides a nice type synonym for plugins that add multiple fields", () => {
-    const plugin: Plugin<{ one: garn.Package; two: garn.Check }> = (p) => ({
-      one: garn.build``,
+    const plugin: garn.Plugin<{ one: garn.Package; two: garn.Check }> = (
+      _p,
+    ) => ({
+      one: garn.build(""),
       two: garn.check(""),
     });
     const project = garn
@@ -169,15 +166,16 @@ describe("Project.add", () => {
   });
 
   it("provides a nice interface for plugins that depend on a non-standard field", () => {
-    const plugin: Plugin<{ addedField: Executable }, { dep: Package }> = (
-      p,
-    ) => ({
+    const plugin: garn.Plugin<
+      { addedField: garn.Executable },
+      { dep: garn.Package }
+    > = (p) => ({
       addedField: garn.shell`${p.dep}/bin/whatever`,
     });
     const project = garn
       .mkProject(
         { description: "", defaultEnvironment: garn.emptyEnvironment },
-        { dep: garn.build`` },
+        { dep: garn.build("") },
       )
       .addExecutable("foo", "")
       .add(plugin);
@@ -188,8 +186,8 @@ describe("Project.add", () => {
   });
 
   it("allows overwriting fields", () => {
-    const plugin: Plugin<{ field: Package }> = (p) => ({
-      field: garn.build``,
+    const plugin: garn.Plugin<{ field: garn.Package }> = (_p) => ({
+      field: garn.build(""),
     });
     const project = garn
       .mkProject(